Integrate DynObject with SimObject (#39)
