+++ pysimenv/core/__init__.py
@@ -2,7 +2,7 @@
 Description for Package
 """
 from pysimenv.core.base import StateVariable, SimObject, StaticObject, BaseFunction
-from pysimenv.core.system import DynObject, DynSystem, TimeVaryingDynSystem
+from pysimenv.core.system import DynSystem, TimeVaryingDynSystem
 from pysimenv.core.simulator import Simulator
 from pysimenv.core.util import SimClock, Timer, Logger
